refactor: player 패키지를 생성하여 플레이어 전용 객체를 player 패키지로 이동
