+++ src/main/java/domain/participant/player/Money.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.participant.player;
 
 import java.math.BigDecimal;
