+++ src/main/java/domain/participant/player/Players.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.participant.player;
 
 import java.util.Collections;
 import java.util.List;
