+++ src/test/java/domain/participant/MoneyTest.java
@@ -1,5 +1,6 @@
 package domain.participant;
 
+import domain.participant.player.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
