+++ src/test/java/domain/participant/PlayerTest.java
@@ -3,6 +3,9 @@ package domain.participant;
 import domain.card.Card;
 import domain.card.Pattern;
 import domain.card.Rank;
+import domain.participant.player.Money;
+import domain.participant.player.Player;
+import domain.participant.player.PlayerInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
